backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
typing
pytest
python-multipart
python-jose 
passlib[bcrypt]==1.7.4
bcrypt==3.2.0